+++ bu/receiver.h
@@ -16,7 +16,7 @@ class Receiver {
 
  public:
   Receiver(std::vector<BuConnectionId> const& connection_ids);
-  size_t receive(double ms_timeout);
+  size_t receive(int ms_timeout);
   size_t receiveAndForget();
 };
